% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMultiAssayExperiment.R
\name{loadMultiAssayExperiment}
\alias{loadMultiAssayExperiment}
\title{Load a MultiAssayExperiment}
\usage{
loadMultiAssayExperiment(
  ds.info,
  project,
  experiments = NULL,
  BPPARAM = NULL,
  include.nested = TRUE
)
}
\arguments{
\item{ds.info}{Named list containing the metadata for this object.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link{acquireFile}}. 
By default, this should be a string containing the path to a staging directory.}

\item{experiments}{Character or integer vector specifying the subset of experiments to load.
If \code{NULL}, all experiments are loaded.}

\item{BPPARAM}{A BiocParallelParam object indicating how loading should be parallelized across multiple experiments.
If \code{NULL}, loading is done serially.}

\item{include.nested}{Logical scalar indicating whether to include nested DataFrames in the \code{colData} of the output.}
}
\value{
A \linkS4class{MultiAssayExperiment} object.
}
\description{
Load a dataset as a \linkS4class{MultiAssayExperiment}, based on the metadata generated by the corresponding \code{\link{stageObject}} method.
}
\examples{
library(SummarizedExperiment)

# Mocking up an MAE
mat <- matrix(rnorm(1000), ncol=10)
colnames(mat) <- letters[1:10]
rownames(mat) <- sprintf("GENE_\%i", seq_len(nrow(mat)))
se <- SummarizedExperiment(list(counts=mat))

library(MultiAssayExperiment)
mae <- MultiAssayExperiment(list(gene=se))

# Staging it:
tmp <- tempfile()
dir.create(tmp)
info <- stageObject(mae, tmp, "dataset")

# Loading it back in:
loadMultiAssayExperiment(info, tmp)

}
\author{
Aaron Lun
}
