% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageMultiAssayExperiment.R
\name{stageObject,MultiAssayExperiment-method}
\alias{stageObject,MultiAssayExperiment-method}
\alias{stageObject,SampleMapFrame-method}
\title{Stage a dataset}
\usage{
\S4method{stageObject}{MultiAssayExperiment}(
  x,
  dir,
  path,
  child = FALSE,
  sm.name = "sample_mapping",
  sd.name = "sample_data",
  meta.name = "dataset.json"
)
}
\arguments{
\item{x}{A \linkS4class{MultiAssayExperiment} object or one of its subclasses.}

\item{dir}{String containing the path to the staging directory.}

\item{path}{String containing a prefix of the relative path inside \code{dir} where \code{x} is to be saved.
The actual path used to save \code{x} may include additional components, see Details.}

\item{child}{Logical scalar indicating whether \code{x} is a child of a larger object.}

\item{sm.name}{String containing the prefix of the sample mapping file.}

\item{sd.name}{String containing the prefix of the sample data file.}

\item{meta.name}{String containing the name of the metadata file.}
}
\value{
A named list containing the metadata for this dataset.
The contents of \code{x} are saved into a \code{path} subdirectory inside \code{dir}.
}
\description{
Save the metadata and annotations of a \linkS4class{MultiAssayExperiment} in a staging directory.
}
\details{
\code{meta.name} is only needed to set up the output \code{path}, for consistency with the \code{\link{stageObject}} contract.
Callers should make sure to write the metadata to the same document by using \code{\link{.writeMetadata}} with \code{meta.only=TRUE}.
}
\examples{
# Mocking up an MAE
mat <- matrix(rnorm(1000), ncol=10)
colnames(mat) <- letters[1:10]
rownames(mat) <- sprintf("GENE_\%i", seq_len(nrow(mat)))

library(SummarizedExperiment)
se <- SummarizedExperiment(list(counts=mat))

library(MultiAssayExperiment)
mae <- MultiAssayExperiment(list(gene=se))

# Staging it:
tmp <- tempfile()
dir.create(tmp)
stageObject(mae, tmp, "dataset")

}
\author{
Aaron Lun
}
